move test to serialization::constant;
